lambda/document_ingest/requirements.txt
boto3
PyPDF2
python-docx
requests
